requirements_test.txt
# WebRTC
selenium
undetected_chromedriver
get-gecko-driver

# GUI
pyautogui
opencv-python

# CorePatched
requests_mock

# TestServer
flask
pyseto

